Min moves to make array increasing
As problem is that we just have to find the minimum moves required to make array increasing, it doesn't say that we have to print that array too. So in case of printing minimum numbers of moves, i came up with an interesting logic but obviously it does not actually print the increasing array .So for the sake of printing array i will do that also but with some other logic obviously.